requirements.txt
aiohttp==3.8.4
discord.py==2.3.0
python-dotenv==1.0.0
torch==2.0.1
torchtext==0.15.2
